Update unit tests
The addition of pipeline/test/test_setup.py was in order to deal with how a bunch
of different classes use ios in various locations. This, compounded by the fact
that ios is an instance, made it exceedingly difficult to mock it out. As a result,
this file will set the output file to a predetermined log file that has been added
to the .gitignore (since writing to /dev/null was not working). This involved
mildly tweaking IOs as well.